changed algorithm drawing BezE, added drag n drop, scaling

+++ font/glyph.h
@@ -10,7 +10,7 @@ class Glyph
 public:
     Glyph();
 
-    int scale;
+    double scale;
     int x;
     int y;
     bool fill;
